[Truffle] Rename to ToEnumSnippets, correct copyright, style fixes
